Move BaseIndex into a separate module.
